Initial import of script for building a package in pbuilder for multiple architectures and distro versions.
This is useful if you want to make backports of a given driver for
people to test with.  It also prints out the paths to the .debs with
build status, so it's easy to fire off the whole build sequence, come
back later, review the status, and upload as appropriate.

Note that there's some constants in the script you'll probably want to
customize if you're looking to use this.  Also, there's tons of room for
improvement and expansion of this script to generalize.  Feel free to
patch it up with enhancements.