auto-xorg-git: fetch branches to local upstream-head-$BRANCH
We used to fetch whatever (-b or not) branch to "upstream-head". This 
would fail if we recycled the git repo, but wanted to fetch another 
branch, since this would also go to "upstream-head". Now we name the
local branch after the remote one, so go happily recycling with all
the branches you want.

To convert an old repo to this format:
 git branch -m upstream-head upstream-head-$BRANCH
where $BRANCH can be found in .git/FETCH_HEAD