auto-xorg-git: automatically force ~ date prefix on new versions
If a .lastversion file contains a given upstream version, a newer 
version can be detected, and will in this case force the -t option date 
prefix to be "~".

This is useful if you use + but want to use ~ next time the upstream 
version is bumped, and you run the script automatically so you might not 
notice the bump before it would be too late.

Also change the default date prefix to be "~" instead of "+", since the 
~ prefix has a few advantages, and reverting from a + to ~ is not 
possible.

So in practise: If you use "+" for a package, write the current upstream 
version into .lastversion.